Combine test files for body matchers
